+++ database/migrations/2023_03_30_074635_create_tempats_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->longText('list_harga');
             $table->longText('keterangan');
             $table->string('gambar')->nullable();
-            $table->integer('jumlah_rating')->nullable();
+            $table->double('jumlah_rating')->nullable();
             $table->timestamps();
         });
     }
